requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
chromadb==1.3.5
streamlit==1.51.0
google-genai==1.52.0